move types back to crate::visitors
